Made changes from usize to u32/u64 wherever appropriate

+++ bench/src/common.rs
@@ -5,7 +5,7 @@ use winter_math::{
     FieldElement,
 };
 
-pub const RUNS: usize = 10;
+pub const RUNS: u32 = 10;
 
 #[derive(ValueEnum, Clone, Debug, Copy)]
 pub enum FieldType {
@@ -15,15 +15,15 @@ pub enum FieldType {
 }
 
 // Parses a string of tuples like "(2,2,0),(4,4,2)" into a Vec of tuples.
-pub fn parse_fri_options(s: &str) -> Result<Vec<(usize, usize, usize)>, String> {
+pub fn parse_fri_options(s: &str) -> Result<Vec<(u32, u32, u32)>, String> {
     s.split(')')
         .filter(|s| !s.trim().is_empty())
         .map(|part| {
             let trimmed = part.trim_start_matches(['(', ',']);
-            let nums: Vec<usize> = trimmed
+            let nums: Vec<u32> = trimmed
                 .split(',')
                 .map(|num_str| num_str.trim().parse())
-                .collect::<Result<Vec<usize>, _>>()
+                .collect::<Result<Vec<u32>, _>>()
                 .map_err(|e| format!("Failed to parse number: {e}"))?;
 
             if nums.len() == 3 {
@@ -38,7 +38,7 @@ pub fn parse_fri_options(s: &str) -> Result<Vec<(usize, usize, usize)>, String> 
         .collect()
 }
 
-pub fn get_standard_fri_options() -> Vec<(usize, usize, usize)> {
+pub fn get_standard_fri_options() -> Vec<(u32, u32, u32)> {
     vec![
         (2, 2, 0),
         (2, 2, 256),
@@ -51,25 +51,25 @@ pub fn get_standard_fri_options() -> Vec<(usize, usize, usize)> {
     ]
 }
 
-pub fn get_standard_data_sizes<E: FieldElement>() -> Vec<usize> {
+pub fn get_standard_data_sizes<E: FieldElement>() -> Vec<u64> {
     vec![
-        (128 * 1024) / E::ELEMENT_BYTES * (E::ELEMENT_BYTES - 1) - 8,
-        (256 * 1024) / E::ELEMENT_BYTES * (E::ELEMENT_BYTES - 1) - 8,
-        (512 * 1024) / E::ELEMENT_BYTES * (E::ELEMENT_BYTES - 1) - 8,
-        (1024 * 1024) / E::ELEMENT_BYTES * (E::ELEMENT_BYTES - 1) - 8,
-        (2048 * 1024) / E::ELEMENT_BYTES * (E::ELEMENT_BYTES - 1) - 8,
+        (128 * 1024) / E::ELEMENT_BYTES as u64 * (E::ELEMENT_BYTES as u64 - 1) - 8,
+        (256 * 1024) / E::ELEMENT_BYTES as u64 * (E::ELEMENT_BYTES as u64 - 1) - 8,
+        (512 * 1024) / E::ELEMENT_BYTES as u64 * (E::ELEMENT_BYTES as u64 - 1) - 8,
+        (1024 * 1024) / E::ELEMENT_BYTES as u64 * (E::ELEMENT_BYTES as u64 - 1) - 8,
+        (2048 * 1024) / E::ELEMENT_BYTES as u64 * (E::ELEMENT_BYTES as u64 - 1) - 8,
     ]
 }
 
-pub fn get_standard_num_queries() -> Vec<usize> {
+pub fn get_standard_num_queries() -> Vec<u32> {
     vec![8, 16, 32]
 }
 
-pub fn get_standard_batch_sizes() -> Vec<usize> {
+pub fn get_standard_batch_sizes() -> Vec<u32> {
     vec![2, 4, 8, 16]
 }
 
-pub fn get_standard_validator_counts() -> Vec<usize> {
+pub fn get_standard_validator_counts() -> Vec<u32> {
     vec![4, 8, 16, 32, 64, 128, 512, 1024]
 }
